backends/drm: fix common mode generation
The unit for refresh rate was wrong, which caused insanely high as well
as negative refresh rates

BUG: 455477
(cherry picked from commit 0c453739b1da4f0e0075fdaf2894e70b17271ef4)

+++ src/backends/drm/drm_object_connector.h
@@ -106,7 +106,7 @@ public:
 
 private:
     QList<QSharedPointer<DrmConnectorMode>> generateCommonModes();
-    QSharedPointer<DrmConnectorMode> generateMode(const QSize &size, uint32_t refreshRate);
+    QSharedPointer<DrmConnectorMode> generateMode(const QSize &size, float refreshRate);
 
     QScopedPointer<DrmPipeline> m_pipeline;
     DrmScopedPointer<drmModeConnector> m_conn;
